fix: storesのshowファイルの、削除ボタンの修正. 管理人にのみ削除ボタンが表示されるようにした.

+++ resources/views/stores/show.blade.php
@@ -28,11 +28,14 @@
                 <a href='/stores/{{ $store->id }}/edit'>[編集]</a>
             </div>
         @endif
-        <form action="/stores/{{ $store->id }}" id="form_{{ $store->id }}" method="post" style="display:inline; padding: 20px 50px;">
-                @csrf
-                @method('DELETE')
-                <button type="submit", onclick="return deleteStore()">[削除]</button> 
-        </form>
+        
+        @if(Auth::id() == implode(config('app.admin')))
+            <form action="/stores/{{ $store->id }}" id="form_{{ $store->id }}" method="post" style="display:inline; padding: 20px 50px;">
+                    @csrf
+                    @method('DELETE')
+                    <button type="submit", onclick="return deleteStore()">[削除]</button> 
+            </form>
+        @endif
 
         <div class='snack' style='padding: 10px 70px;'>
             <h3>お菓子一覧</h3>
